dinsy-client/requirements.txt
configargparse-0.12.0
python-mimeparse-1.6.0
